requirements.txt
# requirements.txt for Meridian Trading System
# Comprehensive requirements based on codebase analysis

# ===== Core Data Processing =====
numpy>=1.21.0                    # Numerical computing
pandas>=1.3.0                    # Data manipulation and analysis
scipy>=1.7.0                     # Scientific computing
numba>=0.56.0                    # JIT compilation for performance

# ===== UI Framework =====
PyQt6>=6.2.0                     # Qt6 Python bindings for GUI
pyqtgraph>=0.12.0                # Fast plotting library for PyQt
qasync>=0.23.0                   # For async QT support

# ===== Web Framework & API =====
fastapi>=0.95.0                  # Modern web framework
uvicorn>=0.20.0                  # ASGI server
pydantic>=1.10.0                 # Data validation

# ===== Database Connectivity =====
supabase>=1.0.0                  # Supabase Python client
postgrest>=0.10.0                # Dependency of supabase
gotrue>=1.0.0                    # Dependency of supabase
realtime>=1.0.0                  # Dependency of supabase
storage3>=0.5.0                  # Dependency of supabase
psutil>=5.9.0                    # System and process monitoring

# ===== HTTP & Networking =====
requests>=2.28.0                 # HTTP library
urllib3>=1.26.0                  # HTTP library (dependency)
aiohttp>=3.8.0                   # Async HTTP client/server
websocket-client>=1.4.0          # For WebSocket connections

# ===== Market Data APIs =====
polygon-api-client>=1.9.0        # Polygon.io API client
yfinance>=0.2.0                  # Yahoo Finance data (if used)

# ===== Data Visualization =====
matplotlib>=3.5.0                # Plotting library
plotly>=5.0.0                    # Interactive plots

# ===== Configuration & Environment =====
python-dotenv>=0.19.0            # Load environment variables from .env
python-dateutil>=2.8.0           # Better date handling
pytz>=2022.1                     # Timezone support

# ===== AI & ML =====
anthropic>=0.25.0                # Anthropic Claude API

# ===== Utilities =====
tabulate>=0.9.0                  # Pretty-print tabular data
orjson>=3.8.0                    # Fast JSON library (used by FastAPI)

# ===== Development & Testing (Optional) =====
# pytest>=7.0.0                  # Testing framework
# black>=22.0.0                  # Code formatter
# flake8>=4.0.0                  # Code linting
# mypy>=0.991                    # Static type checking

# ===== Database Drivers (if needed) =====
# psycopg2-binary>=2.9.0         # PostgreSQL adapter
# sqlalchemy>=1.4.0             # SQL toolkit

# ===== Additional Dependencies =====
# These are standard library modules but including for completeness
# dataclasses (Python 3.7+ built-in)
# typing (Python 3.5+ built-in)
# asyncio (Python 3.4+ built-in)